refactor(web): reorganize app structure and update import paths
- Moved page to a new directory structure under `(main)` for better organization.
- Updated import paths across various components to use absolute imports, enhancing code clarity and maintainability.
- Introduced new components for team management, calendar, and task details, improving the overall functionality of the application.
- Added new pages for user permissions, project management, and timesheet functionalities, expanding the app's capabilities.
