+++ apps/web/app/[locale]/(main)/(teams)/all-teams/component.tsx
