+++ apps/web/app/[locale]/(main)/(teams)/team/[teamId]/[profileLink]/page.tsx
