+++ apps/web/app/[locale]/(main)/(teams)/team/join/page.tsx
