+++ apps/web/app/[locale]/(main)/board/component.tsx
