+++ apps/web/app/[locale]/(main)/board/page.tsx
