+++ apps/web/app/[locale]/(main)/dashboard/team-dashboard/[teamId]/page.tsx
