+++ apps/web/app/[locale]/(main)/integration/github/component.tsx
