+++ apps/web/app/[locale]/(main)/integration/posthog/page-view.tsx
