+++ apps/web/app/[locale]/(main)/kanban/page.tsx
