+++ apps/web/app/[locale]/(main)/meet/jitsi/component.tsx
