+++ apps/web/app/[locale]/(main)/meet/jitsi/page.tsx
