+++ apps/web/app/[locale]/(main)/page.tsx
@@ -1,4 +1,4 @@
-import MainPage from './page-component';
+import MainPage from '../page-component';
 
 // export async function generateStaticParams() {
 // 	return [{ locale: APPLICATION_DEFAULT_LANGUAGE }];
