+++ apps/web/app/[locale]/(main)/permissions/component.tsx
