+++ apps/web/app/[locale]/(main)/permissions/page.tsx
