+++ apps/web/app/[locale]/(main)/projects/[id]/page.tsx
