+++ apps/web/app/[locale]/(main)/settings/layout.tsx
