+++ apps/web/app/[locale]/(main)/task/[id]/component.tsx
