+++ apps/web/app/[locale]/(main)/verify-email/page.tsx
