+++ apps/web/core/components/features/daily-plan/task-details-modal.tsx
@@ -2,7 +2,7 @@ import { Card, Modal } from '@/core/components';
 import { useCallback } from 'react';
 import { clsxm } from '@/core/lib/utils';
 import { ITeamTask } from '@/core/types/interfaces';
-import { TaskDetailsComponent } from '@app/[locale]/task/[id]/component';
+import { TaskDetailsComponent } from '@/app/[locale]/(main)/task/[id]/component';
 import { ScrollArea, ScrollBar } from '@/core/components/ui/scroll-bar';
 
 interface ITaskDetailsModalProps {
